Move getOutcomes and make templateId a number
