+++ app/src/util/networks.ts
@@ -423,3 +423,13 @@ export const getGraphUris = (networkId: number): { httpUri: string; wsUri: strin
   const wsUri = networks[networkId].graphWsUri
   return { httpUri, wsUri }
 }
+
+export const getOutcomes = (networkId: number, templateId: number) => {
+  const isBinary = templateId === 0
+  const isNuancedBinary = (networkId === 1 && templateId === 6) || (networkId === 4 && templateId === 5)
+  if (isBinary || isNuancedBinary) {
+    return ['No', 'Yes']
+  } else {
+    throw new Error(`Cannot get outcomes for network '${networkId}' and template id '${templateId}'`)
+  }
+}
